Fixed unit persistence, add foundations for sound effects/animations, bug fixes
Pretty big commit
Main Changes
- Unit persistence now applies terrain damage when spawning units on the next stage
- Fixed unit collision logic when spawning units on new stage
- Refactored BattleManager and unit scripts to use coroutines in functions that may have
animations or sound effects. This allows the game to wait until these animations are finished
before continuing the game
- BattleManager now waits until units finish moving before continuing
- Refactored the unit selection screen into a prefab. As part of this, the logic for the UI was
moved into its own UnitSelectionWindow script. I also refactored the code so that the window is only
responsible for giving the BattleManager the prefab to spawn.